Updated stream client (#2170)
