+++ demos/common/stream_client/requirements.txt
@@ -1,0 +1,5 @@
+tritonclient[grpc]==2.41.0
+numpy==1.24.3
+ffmpeg-python==0.2.0
+opencv-python==4.9.0.80
+protobuf==4.25.1
