--- demos/real_time_stream_analysis/python/deploy/templates/service.yaml
@@ -1,31 +1,0 @@
-#
-# Copyright (c) 2022 Intel Corporation
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#    http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-{{- if .Values.visualizer_service.port}}
-apiVersion: v1
-kind: Service
-metadata:
-  name: {{ include "rt-stream-analyzer.fullname" . }}
-  labels:
-      app: {{ include "rt-stream-analyzer.fullname" . }}
-spec:
-  type: {{ .Values.visualizer_service.type }}
-  ports:
-    - port: {{ .Values.visualizer_service.port }}
-      targetPort: http
-      protocol: TCP
-      name: http
-{{- end }}
