--- demos/real_time_stream_analysis/python/deploy/values.yaml
@@ -1,41 +1,0 @@
-
-#
-# Copyright (c) 2022 Intel Corporation
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#    http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-image:
-  name: real-time-stream-analysis
-  tag: latest
-  pullPolicy: IfNotPresent
-
-# Set to true if your stream publishing service or OVMS are not deployed in the cluster (but for example as standalone services on your localhost)
-use_host_network: false
-
-visualizer_service:
-  type: ClusterIP
-  # If port set to 0, visualizer will not launch
-  port: 0
-
-stream_url: ""
-ovms_url: ""
-model_name: ""
-model_version: 0
-binary_input: false
-inference_executors: 4
-buffer_size: 100
-log_level: INFO
-
-http_proxy: ""
-https_proxy: ""
