--- demos/real_time_stream_analysis/python/logger.py
@@ -1,39 +1,0 @@
-#
-# Copyright (c) 2022 Intel Corporation
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#    http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-import logging
-
-available_log_levels = ['INFO', 'DEBUG', 'ERROR']
-
-def get_logger(name):
-    logger = logging.getLogger(name)
-    log_formatter = logging.Formatter("%(asctime)s - %(name)s - "
-                                      "%(levelname)s - %(message)s")
-    logger.setLevel(LoggerConfig.log_level)
-    console_handler = logging.StreamHandler()
-    console_handler.setFormatter(log_formatter)
-    logger.addHandler(console_handler)
-
-    return logger
-
-class LoggerConfig:
-    log_level = "INFO"
-
-    @classmethod
-    def set_log_level(cls, log_level: str):
-        log_level = log_level.upper()
-        if log_level in available_log_levels:
-            cls.log_level = log_level
