--- demos/real_time_stream_analysis/python/requirements.txt
@@ -1,3 +1,0 @@
-ovmsclient
-opencv-python
-flask
